Why dictionary in MessageHandler?
If the only thing we are doing with this dictionary is reading values that we defined in the enum class, why use it?
If we need to know what type of messages we are receiving, we can use the composite mark.
Besides... a message does not have to be contained just in the first cell of components!!