Next steps
We just need to write the handlePortMessage method for both MachClient and MachServer.

Notice 1:
The client has to send messages with functionality to the server.
Notice that "definitions.h" holds information about functionality.

Notice 2:
The DataManager does not have names of services, only ports. I think Correspondents do not need anything but ports to which they have a "commitment".

Notice 3:
After this we need to find how to exit loop.

Notice 4:
How do we write the "main"? Maybe the client should not run on a loop??

Notice 5:
We start both using the Correspondent method:

(void) initiateLoopWithDelegate:(id<NSPortDelegate>  _Nullable __strong) delegateObject;

But it now seems to me like we need to do the following:
1. Place the loop in the server.
2. Place the functionality in client.
3. Both handlers should look at the request and behave uniquely to them.

Think about it...